+++ cosmos-model/src/main/scala/com/mesosphere/universe/v3/model/PackageDefinition.scala
@@ -27,11 +27,11 @@ object PackageDefinition {
     }
   }
 
-  final class ReleaseVersion private(val value: Int) extends AnyVal
+  final class ReleaseVersion private(val value: Long) extends AnyVal
 
   object ReleaseVersion {
 
-    def apply(value: Int): Try[ReleaseVersion] = {
+    def apply(value: Long): Try[ReleaseVersion] = {
       if (value >= 0) {
         Return(new ReleaseVersion(value))
       } else {
